tpm2_quote: Enhance autodetect signature support
tpm2_quote by default was hardcoded to:
- TPM2_ALG_RSASSA for RSA keys
- TPM2_ALG_ECDSA  for EC Keys

However, its best to check if the scheme is set by the object for
picking a default, and then defaulting to RSASSA or ECDSA if the
objects scheme in TPM2_ALG_NULL, indicating it allows the caller to
specify the ALG.

If a user specifies an algorithm other than what would be supported
by the object, error out.

Note that ECDAA doesn't work, because you would need to use that with
tpm2_commit to get a valid commit slot to use as the count value.
Currently the tools defaults ECDAA count to 4, per the manpage, but that
doesn't work. Bug #2071 is filled for that.

This will support ECDSA signatures, for example:
  - tpm2_createek -G ecc -c primary.ctx
  - tpm2_createak -C primary.ctx -c ak.ctx -G ecc -g sha256 -s ecdsa
  - tpm2_getrandom -o nonce.bin 20
  - tpm2_quote -c ak.ctx -l sha256:15,16,22 -q nonce.bin -m quote.bin -s quote.sig -o quote.pcr -g sha256

Note that this also changes the default hash size from SHA1 or SHA256
(tool dependent) to SHA256 and updates the manpage. Using SHA1 as the
default was a poor choice, as the whole point of TPM2.0 was SHA256+
algorithm support. This could be seen as a backwards incompatible
change, but users are encouraged to specify the exact algorithm they
want.

Progresses #2068.

Signed-off-by: William Roberts <william.c.roberts@intel.com>