man/tpm2_verifysignature: clarify the scheme option
The -f/scheme option was confusing. One only needs to specify
the --scheme when the non *tss* format is used. The TSS format
contains the metadata identifying the type of signature it is.
It's only the *plain* formats that come from openssl or similair
tools that require the scheme.

Also, that clause said it only supports RSASSA, which is incorrect.

Signed-off-by: William Roberts <william.c.roberts@intel.com>